Break SamSPI into SamSPI, SamSPIInternal, and SamSPIDMA
