+++ MotateProject/motate/Atmel_sam_common/SamSPI.h
@@ -38,211 +38,10 @@
 #include "SamDMA.h"
 #include <type_traits>
 
+#include "SamSPIInternal.h"
+#include "SamSPIDMA.h"
+
 namespace Motate {
-    namespace SPI_internal {
-        template<int8_t spiPeripheralNumber>
-        struct SPIInfo {
-            static constexpr bool exists = false;
-        }; // SPIInfo <generic>
-
-
-        #if defined(SPI)
-            #define CAN_SPI_PDC_DMA 1
-            // We have to strip the address off in static context, outsie a structure declaration
-            static const std::intptr_t _SPI0_address {(std::intptr_t)(SPI)};
-
-            template<>
-            struct SPIInfo<0>
-            {
-                static constexpr bool exists = true;
-                using type = decltype(SPI);
-                static const intptr_t address = _SPI0_address;
-
-                static constexpr RegisterPtr<type> spi { address };
-                static constexpr uint32_t peripheralId = ID_SPI;
-                static constexpr IRQn_Type IRQ = SPI_IRQn;
-            }; // SPIInfo <0>
-            #undef SPI
-            #undef ID_SPI
-            #undef SPI_IRQn
-
-            // This define is because, on this processor, they chose to call it "SPI_Handler" instead
-            // and saves an ifdef and duplicated code in the SamSPI.cpp
-            #define SPI0_Handler SPI_Handler
-        #elif defined(SPI0)
-            // This is for the Sam3x and SamS70
-            // We have to strip the address off in static context, outsie a structure declaration
-            static const std::intptr_t _SPI0_address {(std::intptr_t)(SPI0)};
-
-            template<>
-            struct SPIInfo<0>
-            {
-                static constexpr bool exists = true;
-                static constexpr std::intptr_t address {_SPI0_address};
-                using type = decltype(SPI0);
-
-                static constexpr RegisterPtr<type> spi { address };
-                static constexpr uint32_t peripheralId = ID_SPI0;
-                static constexpr IRQn_Type IRQ = SPI0_IRQn;
-            }; // SPIInfo <0>
-
-            #undef SPI0
-            #undef ID_SPI0
-            #undef SPI0_IRQn
-        #endif
-
-        #if defined(SPI1)
-            // This is for the Sam3x and SamS70
-            // We have to strip the address off in static context, outsie a structure declaration
-            static const std::intptr_t _SPI1_address {(std::intptr_t)(SPI1)};
-
-            template<>
-            struct SPIInfo<1>
-            {
-                static constexpr bool exists = true;
-                static constexpr std::intptr_t address {_SPI1_address};
-                using type = decltype(SPI1);
-
-                static constexpr RegisterPtr<type> spi { address };
-                static constexpr uint32_t peripheralId = ID_SPI1;
-                static constexpr IRQn_Type IRQ = SPI1_IRQn;
-            }; // SPIInfo <0>
-
-            #undef SPI1
-            #undef ID_SPI1
-            #undef SPI1_IRQn
-        #endif
-    }; // Motate::SPI_internal
-
-    // We're deducing if there's a SPI0 and it has a PDC
-    #if defined(CAN_SPI_PDC_DMA)
-
-    #else
-    #pragma mark DMA_XDMAC SPI implementation
-
-    template<uint8_t spiPeripheralNumber>
-    struct DMA_XDMAC_hardware<Spi*, spiPeripheralNumber> : Motate::SPI_internal::SPIInfo<spiPeripheralNumber>
-    {
-        using SPI_internal::SPIInfo<spiPeripheralNumber>::spi;
-        static constexpr auto peripheralId = SPI_internal::SPIInfo<spiPeripheralNumber>::peripheralId;
-        typedef char* buffer_t;
-    };
-
-    template<uint8_t spiPeripheralNumber>
-    struct DMA_XDMAC_TX_hardware<Spi*, spiPeripheralNumber> : virtual DMA_XDMAC_hardware<Spi*, spiPeripheralNumber>, virtual DMA_XDMAC_common {
-        using DMA_XDMAC_hardware<Spi*, spiPeripheralNumber>::spi;
-        using DMA_XDMAC_hardware<Spi*, spiPeripheralNumber>::peripheralId;
-
-        static constexpr uint8_t const xdmaTxPeripheralId()
-        {
-            switch (spiPeripheralNumber) {
-                case (0): return 1;
-                case (1): return 3;
-            };
-            return 0;
-        };
-        static constexpr uint8_t const xdmaTxChannelNumber()
-        {
-            switch (spiPeripheralNumber) {
-                case (0): return  18;
-                case (1): return  20;
-            };
-            return 0;
-        };
-        static constexpr XdmacChid * const xdmaTxChannel()
-        {
-            return xdma()->XDMAC_CHID + xdmaTxChannelNumber();
-        };
-        static constexpr volatile void * const xdmaPeripheralTxAddress()
-        {
-            return &spi->SPI_TDR;
-        };
-    };
-
-    template<uint8_t spiPeripheralNumber>
-    struct DMA_XDMAC_RX_hardware<Spi*, spiPeripheralNumber> : virtual DMA_XDMAC_hardware<Spi*, spiPeripheralNumber>, virtual DMA_XDMAC_common {
-        using DMA_XDMAC_hardware<Spi*, spiPeripheralNumber>::spi;
-        using DMA_XDMAC_hardware<Spi*, spiPeripheralNumber>::peripheralId;
-
-        static constexpr uint8_t const xdmaRxPeripheralId()
-        {
-            switch (spiPeripheralNumber) {
-                case (0): return 2;
-                case (1): return 4;
-            };
-            return 0;
-        };
-        static constexpr uint8_t const xdmaRxChannelNumber()
-        {
-            switch (spiPeripheralNumber) {
-                case (0): return 19;
-                case (1): return 21;
-            };
-            return 0;
-        };
-        static constexpr XdmacChid * const xdmaRxChannel()
-        {
-            return xdma()->XDMAC_CHID + xdmaRxChannelNumber();
-        };
-        static constexpr volatile void * const xdmaPeripheralRxAddress()
-        {
-            return &spi->SPI_RDR;
-        };
-    };
-
-    // Construct a DMA specialization that uses the PDC
-    template<uint8_t periph_num>
-    struct DMA<Spi*, periph_num> : DMA_XDMAC_RX<Spi*, periph_num>, DMA_XDMAC_TX<Spi*, periph_num> {
-        // nothing to do here, except for a constxpr constructor
-        constexpr DMA(const std::function<void(uint16_t)> &handler) : DMA_XDMAC_RX<Spi*, periph_num>{handler}, DMA_XDMAC_TX<Spi*, periph_num>{handler} {};
-
-        using DMA_XDMAC_common::xdma;
-        using DMA_XDMAC_RX<Spi*, periph_num>::xdmaRxChannelNumber;
-        using DMA_XDMAC_TX<Spi*, periph_num>::xdmaTxChannelNumber;
-
-        void setInterrupts(const uint16_t interrupts) const
-        {
-            DMA_XDMAC_common::setInterrupts(interrupts);
-
-            if (interrupts != Interrupt::Off) {
-                if (interrupts & Interrupt::OnTxTransferDone) {
-                    DMA_XDMAC_TX<Spi*, periph_num>::startTxDoneInterrupts();
-                } else {
-                    DMA_XDMAC_TX<Spi*, periph_num>::stopTxDoneInterrupts();
-                }
-
-                if (interrupts & Interrupt::OnRxTransferDone) {
-                    DMA_XDMAC_RX<Spi*, periph_num>::startRxDoneInterrupts();
-                } else {
-                    DMA_XDMAC_RX<Spi*, periph_num>::stopRxDoneInterrupts();
-                }
-            } else {
-                DMA_XDMAC_TX<Spi*, periph_num>::stopTxDoneInterrupts();
-                DMA_XDMAC_RX<Spi*, periph_num>::stopRxDoneInterrupts();
-
-            }
-        };
-
-        void reset() const {
-            DMA_XDMAC_TX<Spi*, periph_num>::resetTX();
-            DMA_XDMAC_RX<Spi*, periph_num>::resetRX();
-        };
-
-        void enable() const {
-            xdma()->XDMAC_GE = (XDMAC_GIE_IE0 << xdmaRxChannelNumber()) | (XDMAC_GIE_IE0 << xdmaTxChannelNumber());
-        }
-
-        void disable() const
-        {
-            xdma()->XDMAC_GD = (XDMAC_GID_ID0 << xdmaRxChannelNumber()) | (XDMAC_GID_ID0 << xdmaRxChannelNumber());
-        };
-    };
-    #endif // SPI + XDMAC
-
-    #undef HAS_SPI1
-    #undef HAS_SPI0
-
-
     template<int8_t spiPeripheralNumber>
     struct _SPIHardware : Motate::SPI_internal::SPIInfo<spiPeripheralNumber>
     {
